+++ src/dummy.js
@@ -1,0 +1,2 @@
+/* this file is just a dummy */
+export const DUMMY_PATH = __filename;