+++ src/entry-loader.js
@@ -1,0 +1,6 @@
+import loaderUtils from 'loader-utils';
+
+module.exports = function () {
+   const { files } = loaderUtils.getOptions(this);
+   return files.map( file => `require( '${file}' )` ).join('\n');
+};