+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Last Changes
 
+- [#5](https://github.com/LaxarJS/laxar-infrastructure/issues/5): update karma setup so it can handle multiple specs
+- [#4](https://github.com/LaxarJS/laxar-infrastructure/issues/4): fix browser-spec / webpack-dev-server usage
+
 
 ## v0.1.0-alpha.1